+++ projects/llm_d/orchestration/runtime_config.py
@@ -1,16 +1,11 @@
 from __future__ import annotations
 
 import copy
-import hashlib
-import json
 import logging
 import re
 from pathlib import Path
 from typing import Any
 
-import yaml
-
-from projects.core.dsl.utils import slugify_identifier, truncate_k8s_name
 from projects.core.library import config, env, run
 
 logger = logging.getLogger(__name__)
@@ -41,11 +36,6 @@ def ensure_artifact_directories(artifact_dir: Path) -> None:
 # Configuration accessor functions
 
 
-def get_artifact_dir() -> Path:
-    """Get the current artifact directory"""
-    return Path(env.ARTIFACT_DIR)
-
-
 def get_config_dir() -> Path:
     """Get the LLM-D configuration directory"""
     return ORCHESTRATION_DIR
@@ -53,8 +43,6 @@ def get_config_dir() -> Path:
 
 def get_namespace() -> str:
     """Get the resolved namespace for this execution"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
 
     platform_data = get_platform_config()
     namespace_override = config.project.get_config("runtime.namespace_override", None)
@@ -77,8 +65,6 @@ def get_namespace() -> str:
 
 def get_namespace_is_managed() -> bool:
     """Check if namespace is managed (auto-derived) vs explicitly configured"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
 
     namespace_override = config.project.get_config("runtime.namespace_override", None)
     platform_data = get_platform_config()
@@ -89,8 +75,6 @@ def get_namespace_is_managed() -> bool:
 
 def get_job_name() -> str:
     """Get the resolved job name"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
 
     job_name = config.project.get_config("runtime.job_name", None)
     if job_name:
@@ -102,37 +86,32 @@ def get_job_name() -> str:
 
 def get_model_key() -> str:
     """Get the selected model key"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
+
     return config.project.get_config("runtime.model_key")
 
 
 def get_model() -> dict[str, Any]:
     """Get the resolved model configuration"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
+
     model_key = get_model_key()
     return copy.deepcopy(config.project.get_config(f"models.{model_key}"))
 
 
 def get_platform_config() -> dict[str, Any]:
     """Get the normalized platform configuration"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
+
     return normalize_platform_config(copy.deepcopy(config.project.get_config("platform")))
 
 
 def get_model_cache_config() -> dict[str, Any]:
     """Get the model cache configuration"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
+
     return copy.deepcopy(config.project.get_config("model_cache"))
 
 
 def get_benchmark_config() -> dict[str, Any] | None:
     """Get the benchmark configuration if specified"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
+
     benchmark_name = config.project.get_config("runtime.benchmark_key", None)
     if not benchmark_name:
         return None
@@ -141,8 +120,7 @@ def get_benchmark_config() -> dict[str, Any] | None:
 
 def get_scheduler_profile() -> dict[str, Any] | None:
     """Get the scheduler profile configuration"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
+
     profile_key = config.project.get_config("runtime.scheduler_profile_key")
     if profile_key == "default":
         return None
@@ -151,22 +129,19 @@ def get_scheduler_profile() -> dict[str, Any] | None:
 
 def get_gpu_count() -> int | None:
     """Get the normalized GPU count"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
+
     return normalize_gpu_count(config.project.get_config("runtime.gpu_count", None))
 
 
 def get_scheduler_profile_key() -> str:
     """Get the selected scheduler profile key"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
+
     return config.project.get_config("runtime.scheduler_profile_key")
 
 
 def get_smoke_request() -> dict[str, Any]:
     """Get the smoke request configuration"""
-    if config.project is None:
-        raise RuntimeError("Configuration not initialized")
+
     smoke_request_key = config.project.get_config("runtime.smoke_request_key")
     return copy.deepcopy(config.project.get_config(f"workloads.smoke_requests.{smoke_request_key}"))
 
@@ -222,87 +197,3 @@ def derive_namespace(job_name: str, prefix: str, max_length: int) -> str:
 def version_tuple(value: str) -> tuple[int, ...]:
     numbers = re.findall(r"\d+", value)
     return tuple(int(number) for number in numbers[:3])
-
-
-def resolve_model_cache_spec(
-    *,
-    namespace: str,
-    model_key: str,
-    model: dict[str, Any],
-    model_cache: dict[str, Any],
-) -> dict[str, Any] | None:
-    if not model_cache.get("enabled", False):
-        return None
-
-    source_uri = model["uri"]
-    if source_uri.startswith(("pvc://", "pvc+hf://")):
-        return None
-
-    if source_uri.startswith("hf://"):
-        source_scheme = "hf"
-    elif source_uri.startswith("oci://"):
-        source_scheme = "oci"
-    else:
-        raise ValueError(f"Unsupported model cache source URI for {model_key}: {source_uri}")
-
-    model_cache_overrides = model.get("cache", {})
-    pvc_defaults = model_cache["pvc"]
-    pvc_prefix = model_cache["pvc"]["name_prefix"]
-    cache_key = hashlib.sha256(source_uri.encode("utf-8")).hexdigest()[:10]
-    pvc_name = truncate_k8s_name(
-        f"{pvc_prefix}-{slugify_identifier(model_key, max_length=32)}-{cache_key}"
-    )
-    model_path = pvc_defaults["model_directory_name"]
-
-    return {
-        "source_uri": source_uri,
-        "source_scheme": source_scheme,
-        "cache_key": cache_key,
-        "namespace": namespace,
-        "pvc_name": pvc_name,
-        "pvc_size": model_cache_overrides.get("pvc_size", pvc_defaults["size"]),
-        "access_mode": model_cache_overrides.get("access_mode", pvc_defaults["access_mode"]),
-        "storage_class_name": model_cache_overrides.get(
-            "storage_class_name", pvc_defaults.get("storage_class_name")
-        ),
-        "model_path": model_path,
-        "model_uri": f"pvc://{pvc_name}/{model_path}",
-        "marker_filename": model_cache["marker_filename"],
-        "marker_path": f"/cache/{model_path}/{model_cache['marker_filename']}",
-        "download_job_name": truncate_k8s_name(f"{pvc_name}-download"),
-        "hf_token_secret_name": model_cache_overrides.get(
-            "hf_token_secret_name", model_cache["hf"].get("token_secret_name")
-        ),
-        "hf_token_secret_key": model_cache["hf"].get("token_secret_key"),
-        "oci_image_path": model_cache_overrides.get(
-            "oci_image_path", model_cache["oci"].get("image_path")
-        ),
-        "oci_registry_auth_secret_name": model_cache_overrides.get(
-            "oci_registry_auth_secret_name",
-            model_cache["oci"].get("registry_auth_secret_name"),
-        ),
-        "oci_registry_auth_secret_key": model_cache["oci"].get("registry_auth_secret_key"),
-    }
-
-
-def load_yaml(path: Path) -> Any:
-    with path.open(encoding="utf-8") as handle:
-        return yaml.safe_load(handle)
-
-
-def write_yaml(path: Path, payload: Any) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    with path.open("w", encoding="utf-8") as handle:
-        yaml.safe_dump(payload, handle, sort_keys=False)
-
-
-def write_json(path: Path, payload: Any) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    with path.open("w", encoding="utf-8") as handle:
-        json.dump(payload, handle, indent=2, sort_keys=True)
-        handle.write("\n")
-
-
-def write_text(path: Path, content: str) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(content, encoding="utf-8")
